arm64: dts: qcom: msm8916-samsung-a5ltezt: Add initial device tree (v2)
Samsung Galaxy A5 (SM-A500YZ) is a smartphone using the MSM8916 SoC released
in 2015.

The A5 variants are very similar, with some differences in accelerometer and
touch key, yet a5ltezt doesn't have an NFC chip.
The common parts are shared in msm8916-samsung-a5u-eur.dts to reduce
duplication.

The difference of accelerometer between a5ulte (a5u-eur) / a5ltezt is the way
that the sensor is mounted on the mainboard - set the mount-matrix in the
device-specific device tree part to handle that difference.

On a5ltezt the touch key is supplied by a single fixed
regulator (enabled via GPIO 60) that supplies both MCU and LED.

v2: rename a5-zt to a5ltezt
Signed-off-by: "Lin, Meng-Bo" <linmengbo06890@proton.me>