arm64: dts: qcom: msm8916-samsung-heatqlte: Add initial device tree
Samsung Galaxy Ace 4 is a phone based on MSM8916. It is similar
to the other Samsung devices based on MSM8916 with only a few minor
differences.

Ace 4 and Core Prime are similar, with some differences in panel and
touchscreen. The common parts are shared in
msm8916-samsung-rossa-common.dtsi to reduce duplication.

Unfortunately, SM-G357FZ was released with outdated 32-bit only
firmware and never received any update from Samsung. Since the 32-bit
TrustZone firmware is signed there seems to be no way currently to
actually boot this device tree on arm64 Linux at the moment.

However, it is possible to use this device tree by compiling an ARM32
kernel instead. The device tree can be easily built on ARM32 with
an #include and it works really well there. To avoid confusion for others
it is still better to add this device tree on arm64. Otherwise it's easy
to forget to update this one when making some changes that affect all
MSM8916 devices.

Maybe someone finds a way to boot ARM64 Linux on this device at some
point.
In this case we expect that this device tree can be simply used as-is.

Signed-off-by: Gareth Peoples <mail@gpeopl.es>
[Raymond: Use msm8916-samsung-rossa-common.dtsi and reword the commit]
v2: refactor MUIC and touchscreen
Signed-off-by: Raymond Hackley <raymondhackley@protonmail.com>